Fix user interaction config handling in image & directory install modes (#1379106)
Don't write out the user interaction config file for image and directory
installation modules and don't read the user configuration file from
the host.

In these modes Anaconda actually prepares the installation image,
so storing the config file could result in spokes being hidden
during the actual installation. Reading the configuration
file from the host could also cause havoc - by hiding random
spoke based on what spokes the user visited when installing the
host system.

For example if the storage spokes was visited during an interactive
image or directory installation, it would not be shown during the
actual installation run from the generated image/directory,
as it would be considered already visited. If the user configuration
file from the host was read, it would also cause all spokes
visited during the *host installation* to disappear.

So better not do that. :)